+++ NEWS.md
@@ -1,5 +1,12 @@
 # QFeatures 1.13
 
+## QFeatures 1.13.3
+
+- `readQFeatures()` multi-set support (ported from `scp::readSCP()` -
+  see issue #199).
+- new `readQFeaturesFromDIANN()` function to import DIA-NN report
+  files (ported from `scp::readSCPfromDIANN()` - see issue #199).
+
 ## QFeatures 1.13.2
 
 - Move the `filterFeatures` generic method to `ProtGenerics`.